Add xr input device controller and currently used controller
The XRInputDeviceController class now handles the input of the InputDevices and no longer the InputManager which now manages the XRInputDeviceController instances.
This was done to add the support of adressing the CurrentlyUsedController instead of only being able to subscribe to events of the left or right controller.
Usages have already been adjusted.